% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc.LCMV}
\alias{cc.LCMV}
\title{cc.LCMV Data}
\format{
A list with 12 elements (one per modulon) containing all the connected components of the modulon
}
\source{
\url{https://github.com/icrespocasajus/ModulonCore}
}
\usage{
cc.LCMV
}
\description{
A dataset containing the connected components of each modulon.
}
\examples{
cc.LCMV
}
\keyword{datasets}
