% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DA.LCMV}
\alias{DA.LCMV}
\title{DA.LCMV Data}
\format{
A list of 6 elements.
}
\usage{
DA.LCMV
}
\description{
A dataset containing the OPLS-DA analysis score (WeightStarMN) normalized.
}
\examples{
DA.LCMV
}
\keyword{datasets}
