% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DA.TILs}
\alias{DA.TILs}
\title{DA.TILs Data}
\format{
A list of 5 elements.
}
\usage{
DA.TILs
}
\description{
A dataset containing the OPLS-DA analysis score (WeightStarMN) normalized.
}
\examples{
DA.TILs
}
\keyword{datasets}
