% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{Filter.Sat}
\alias{Filter.Sat}
\title{Filter satellite transcription factors by discriminant power}
\usage{
Filter.Sat(sat.data, DA.data, DA = c("Any"), top.percent = 10)
}
\arguments{
\item{sat.data}{List object with as many elements as the modulon connected components provided as the input. Each element of the list contains a character vector with satellite transcription factors.}

\item{DA.data}{List object with the discriminant score of all the regulons derived from one or more discriminant analysis.}

\item{DA}{OPLS-DA to be considered. This argument can take as value any of the names of DA.data list or several of them. If DA = 'Any', all the names of DA.data are considered.}

\item{top.percent}{Quantile (in \%) to be considered from the top of the discriminant score ranking.}
}
\value{
List object with as many elements as the modulon connected components provided in the input. Each element of the list contains a character vector with the corresponding satellite transcription factors that are among the top discriminant regulons from one or more discriminant analysis.
}
\description{
Filter satellite transcription factors by discriminant power values derived from one or more discriminant analysis.
}
\details{
This function remove satellite transcription factors with low discriminant power.
}
\examples{
\dontrun{
if(interactive()){
target.analysis.modulon.wrt.cc.manual.query(net = network.TILs,
mod = modulons.TILs,
cc = cc.TILs,
query.mod = '3',
query.cc= 'cc.3') \%>\% Find.Sat() \%>\% Filter.Sat(DA.data = DA.TILs)
}
}
}
