% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{jaccard}
\alias{jaccard}
\title{Jaccard Distance}
\usage{
jaccard(a, b)
}
\arguments{
\item{a}{A vector.}

\item{b}{A vector.}
}
\value{
Jaccard distance calculated as the ratio between the intersection and the union of the two vectors.
}
\description{
Calculate Jaccard distance between two vectors.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 jaccard(c('A','B','C','D','E'), c('A','B','C'))
 }
}
}
