% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Modulon.Cores.TILs}
\alias{Modulon.Cores.TILs}
\title{Modulon.Cores.TILs Data}
\format{
A character vector of 5 elements.
}
\source{
\url{https://github.com/icrespocasajus/ModulonCore}
}
\usage{
Modulon.Cores.TILs
}
\description{
A dataset containing the names of the modulon regulatory cores.
}
\examples{
Modulon.Cores.TILs
}
\keyword{datasets}
