% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{Modulon.corrplot}
\alias{Modulon.corrplot}
\title{Plot modulon target similarity, redundancy or overlap.}
\usage{
Modulon.corrplot(net, mod, cc, regulatory.core, feature = "Redundancy", RegAUC)
}
\arguments{
\item{net}{List R object with as many elements as the modulon connected components provided as the input. Each element of the list contains a character vector with satellite transcription factors.}

\item{mod}{List object with each modulon constituent elements.}

\item{cc}{List with the connected components generated with find.connected.components() or regulatory cores as the output of core()}

\item{regulatory.core}{Regulatory core names.}

\item{feature}{Target analysis feature to be displayed; one of c("Redundancy","Similarity","Overlap")}

\item{RegAUC}{Regulon activity matrix.}
}
\value{
List object with as many elements as modulons; each element contain a heatmap.
}
\description{
Plot a heatmap displaying modulon target similarity, redundancy or overlap.
}
\details{
This function generates a heatmap to explore the results of the modulon target analysis
}
\examples{
\dontrun{
if(interactive()){
plots = Modulon.corrplot(
 net = network.TILs,
 mod = modulons.TILs,
 cc = cc.TILs,
 regulatory.core = Modulon.Cores.TILs,
 feature = 'Redundancy',
 RegAUC = RegAUC.TILs)
 print(plots[['3']])
}
}
}
