% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{Modulon.Membership}
\alias{Modulon.Membership}
\title{Calculate modulon membership}
\usage{
Modulon.Membership(data, mod)
}
\arguments{
\item{data}{Matrix of dataframe with regulon activity.}

\item{mod}{List object with each modulon constituent elements.}
}
\value{
List object with as many elements as modulons. Each element of the list contains a character vector with the corresponding modulon constituent elements.
}
\description{
Calculation of the correlation (Spearman) between the first principal component (PC1) of a subset of the regulon activity matrix with only modulon constituent elements and each of these elements.
}
\details{
This function subsets the regulon activity matrix to include only modulon constituent elements. After centering and scaling the resulting matrix, a PCA is performed; the PC1 is then correlated with the regulon activity of each modulon constituent element.
}
\examples{
\dontrun{
if(interactive()){
Modulon.Membership(data = RegAUC.TILs, mod = modulons.TILs)
}
}
}
