% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modulons.LCMV}
\alias{modulons.LCMV}
\title{modulons.LCMV Data}
\format{
A list of 12 elements/modulons with the modulon constituent elements
}
\source{
\url{https://www.nature.com/articles/s41467-021-23324-4}
}
\usage{
modulons.LCMV
}
\description{
A dataset containing the modulons inferred using modulon analysis on murine LCMV transcriptomics.
}
\examples{
modulons.LCMV
}
\keyword{datasets}
