% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{modulons.TILs}
\alias{modulons.TILs}
\title{modulons.TILs Data}
\format{
A list of 7 elements/modulons with the modulon constituent elements
}
\source{
\url{https://www.nature.com/articles/s41467-021-23324-4}
}
\usage{
modulons.TILs
}
\description{
A dataset containing the modulons inferred using modulon analysis on murine TILs transcriptomics.
}
\examples{
modulons.TILs
}
\keyword{datasets}
