% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{network.TILs}
\alias{network.TILs}
\title{network.TILs Data}
\format{
A dataframe encoding the network in 3 columns:'Source','Interaction','Target'
}
\source{
\url{https://www.nature.com/articles/s41467-021-23324-4}
}
\usage{
network.TILs
}
\description{
A dataset containing the transcriptional network inferred using SCENIC pipeline on murine TILs transcriptomics.
}
\examples{
network.TILs
}
\keyword{datasets}
