% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{overlap}
\alias{overlap}
\title{Overlap between two vectors
Calculate the number of common elements between two vectors}
\usage{
overlap(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
The integer with the number of common elements
}
\description{
Overlap between two vectors
Calculate the number of common elements between two vectors
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
overlap(c('A','B','C'),c('A','B','C','D','E'))
 }
}
}
