% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{range01}
\alias{range01}
\title{Range between 0 and 1}
\usage{
range01(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
Normalized values (x-min(x))/(max(x)-min(x))
}
\description{
Normalize absolute values of a numeric vector
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
range01(c('-50','-5','5','50','100')
 }
}
}
