% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{redundancy}
\alias{redundancy}
\title{Redundancy
Calculate redundancy between two vectors}
\usage{
redundancy(a, b)
}
\arguments{
\item{a}{A vector}

\item{b}{A vector}
}
\value{
Redundancy calculated as the ratio between the intersection and the minimum length of the two vectors.
}
\description{
Redundancy
Calculate redundancy between two vectors
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
redundancy(c('A','B','C','D','E'), c('A','B','C'))
 }
}
}
