% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegAUC.LCMV}
\alias{RegAUC.LCMV}
\title{RegAUC.LCMV Data}
\format{
A matrix with 4000 rows (cells) and 646 columns (regulons).
}
\usage{
RegAUC.LCMV
}
\description{
Regulon AUC values
}
\examples{
RegAUC.LCMV
}
\keyword{datasets}
