% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RegAUC.TILs}
\alias{RegAUC.TILs}
\title{RegAUC.TILs Data}
\format{
A matrix with 11379 rows (cells) and 583 columns (regulons).
}
\usage{
RegAUC.TILs
}
\description{
Regulon AUC values
}
\examples{
RegAUC.TILs
}
\keyword{datasets}
