% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{strsplit2}
\alias{strsplit2}
\title{Split Composite Names}
\usage{
strsplit2(x, split, ...)
}
\arguments{
\item{x}{character vector}

\item{split}{character to split each element of vector on, see strsplit}

\item{...}{other arguments are passed to strsplit}
}
\value{
Normalized values (x-min(x))/(max(x)-min(x))
}
\description{
Split a vector of composite names into a matrix of simple names.
}
\details{
This function is the same as strsplit except that the output value is a matrix instead of a list.
The first column of the matrix contains the first component from each element of x, the second column contains the second components etc.
The number of columns is equal to the maximum number of components for any element of x.
The motivation for this function in the limma package is handle input columns which are composites of two or more annotation fields.
}
\examples{
\dontrun{
if(interactive()){
x = c("2__cc.1","2__cc.2"); strsplit2(x,split="__")
 }
}
}
