% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{target.analysis.manual.query}
\alias{target.analysis.manual.query}
\title{Target Analysis manual query}
\usage{
target.analysis.manual.query(net, modulon, cc, query.mod, query.cc)
}
\arguments{
\item{net}{A dataframe with a network encoded in 3 columns: 'Source','Interaction','Target'.}

\item{cc}{Connected components generated with find.connected.components() or regulatory cores as the output of core()}

\item{query.mod}{Name of the query modulon}

\item{query.cc}{Name of the connected component}

\item{modulons}{A list with as many elements as modulons/clusters containing the constituent elements.}
}
\value{
List with as many elements as the modulon constituent elements out of a given modulon connected component. Each element includes a dataframe with 4 columns: TF, Redundancy, Similarity and Overlap
}
\description{
Find common targets between a given modulon connected component and modulon constituent elements out of the connected component; the specific modulon and connected component have to be specified.
}
\details{
This function is searches for the modulon constituent elements sharing targets with a given modulon connected component.
}
\examples{
\dontrun{
if(interactive()){
target.analysis.manual.query(net = network.TILs,
mod = modulons.TILs,
cc = cc.TILs,
query.mod = '3',
query.cc= 'cc.3')
}
}
}
