% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{target.analysis.modulon.plot}
\alias{target.analysis.modulon.plot}
\title{Plot results of the modulon target analysis}
\usage{
target.analysis.modulon.plot(
  data,
  feature = "Redundancy",
  color = "YlGn",
  core = c(),
  core.sat = c()
)
}
\arguments{
\item{data}{List object with three elements ("Redundancy","Similarity" and "Overlap") with the corresponding matrices.}

\item{feature}{Target analysis feature to be displayed; one of c("Redundancy","Similarity","Overlap").}

\item{core}{Character vector with modulon regulatory core constituent elements.}

\item{core.sat}{Character vector with modulon regulatory core detected satellites.}
}
\value{
A corrplot() object.
}
\description{
Generate the plot for the matrix with all pair-wise comparisons within a given modulon for one of the following features:"Redundancy","Similarity" and "Overlap".
}
\details{
This function generates a plot showing the results of the modulon target analysis of a given modulon.
}
\examples{
\dontrun{
if(interactive()){
results.target.analysis.modulon=target.analysis.modulon(net=network.TILs,mod=modulons.TILs,mod.query = '3')
target.analysis.modulon.plot(data=results.target.analysis.modulon,feature = 'Redundancy')
}
}
}
