% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{target.analysis.modulon}
\alias{target.analysis.modulon}
\title{Modulon target analysis: similarity, redundancy and overlap}
\usage{
target.analysis.modulon(net, mod, mod.query)
}
\arguments{
\item{net}{List R object with as many elements as the modulon connected components provided as the input. Each element of the list contains a character vector with satellite transcription factors.}

\item{mod}{List object with each modulon constituent elements.}

\item{mod.query}{The name of the modulon to be analysed.}
}
\value{
List object with three objects ("Redundancy","Similarity" and "Overlap") with the corresponding matrices.
}
\description{
Generate the matrix with all pair-wise comparisons within a given modulon for each of the following features:"Redundancy","Similarity" and "Overlap".
}
\details{
This function perform the modulon target analysis.
}
\examples{
\dontrun{
if(interactive()){
target.analysis.modulon(net=network.TILs,mod=modulons.TILs,mod.query = '3')
}
}
}
