% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA_functions.R
\name{target.analysis.modulon.wrt.cc.manual.query}
\alias{target.analysis.modulon.wrt.cc.manual.query}
\title{Target Analysis with respect to modulon connected components: manual query}
\usage{
target.analysis.modulon.wrt.cc.manual.query(net, mod, cc, mod.query, cc.query)
}
\arguments{
\item{net}{A dataframe with a network encoded in 3 columns: 'Source','Interaction','Target'.}

\item{mod}{A list with as many elements as modulons/clusters containing the constituent elements.}

\item{cc}{Connected components generated with find.connected.components() or regulatory cores as the output of core().}

\item{mod.query}{Name of the query modulon.}

\item{cc.query}{Name of the connected component.}
}
\value{
List with as many elements as the modulon constituent elements out of a given modulon connected component. Each element includes a dataframe with 4 columns: TF, Redundancy, Similarity and Overlap
}
\description{
Find common targets between a given modulon connected component and modulon constituent elements out of the connected component; the specific modulon and connected component have to be specified.
}
\details{
This function is searches for the modulon constituent elements sharing targets with a given modulon connected component.
}
\examples{
\dontrun{
if(interactive()){
target.analysis.modulon.wrt.cc.manual.query(net = network.TILs,
mod = modulons.TILs,
cc = cc.TILs,
mod.query = '3',
cc.query = 'cc.3')
}
}
}
