% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModulonTA-package.R
\docType{package}
\name{ModulonTA-package}
\alias{ModulonTA}
\alias{ModulonTA-package}
\title{ModulonTA: Modulon Target Analysis}
\description{
ModulonTA performs a target analysis on modulon constituent elements.
}
\author{
\strong{Maintainer}: Isaac Crespo \email{isaaccrespo@hotmail.com} (\href{https://orcid.org/0000-0003-3151-3901}{ORCID})

}
\keyword{internal}
